NODE-1161: restore timestamp

+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/traits/Domain.scala
@@ -33,7 +33,7 @@ object DataItem {
   case class Str(k: String, v: String)     extends DataItem[String]     { val key = k; val value = v }
 }
 
-case class Blk(timestamp: Long, height: Int, generationSignature: ByteVector)
+case class Blk(timestamp: Long, height: Long, generationSignature: ByteVector)
 
 case class Ord(id: ByteVector,
                sender: Recipient.Address,
